исправил замечания по homework-3
